ttw_backend/api/serializers.py
from rest_framework import serializers
from django.contrib.auth import authenticate
from .models import CustomUser

class SignupSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True, required=True, min_length=6)

    class Meta:
        model = CustomUser
        fields = ('username', 'password')

    def validate_username(self, value):
        if CustomUser.objects.filter(username=value).exists():  # Change to 'email' if renamed
            raise serializers.ValidationError("Username is already taken.")
        return value

    def create(self, validated_data):
        user = CustomUser.objects.create_user(
            username=validated_data['username'],  # Change to 'email' if renamed
            password=validated_data['password']
        )
        return user

class LoginSerializer(serializers.Serializer):
    username = serializers.CharField()
    password = serializers.CharField(write_only=True)

    def validate(self, data):
        user = authenticate(username=data['username'], password=data['password'])
        if user and user.is_active:
            return user
        raise serializers.ValidationError("Invalid credentials.")
