requirements.txt
ansicon==1.89.0
anyio==4.7.0
arrow==1.3.0
asgiref==3.8.1
autopep8==2.3.1
black==24.10.0
blessed==1.20.0
cfgv==3.4.0
channels==4.2.0
channels_redis==4.2.1
click==8.1.7
colorama==0.4.6
distlib==0.3.9
Django==5.1.4
django-cors-headers==4.6.0
django-picklefield==3.2
django-q2==1.7.4
djangorestframework==3.15.2
filelock==3.16.1
flake8==7.1.1
h11==0.14.0
httptools==0.6.4
identify==2.6.3
idna==3.10
isort==5.13.2
jinxed==1.3.0
mccabe==0.7.0
msgpack==1.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
virtualenv==20.28.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
